fix issue with handling new files where decoration is requested before file watcher handles change, updating caching, and improve logging

+++ src/classes/LinkManager.ts
@@ -17,7 +17,7 @@ type PathData = {
 };
 
 export default class LinkManager {
-  readonly #pathDataCache: Map<NormalisedPath, PathData | null> = new Map();
+  private readonly pathDataCache: Map<NormalisedPath, PathData | null> = new Map();
 
   /**
    * This represents all the relevant paths the link manager is aware of that are of a known type
@@ -40,7 +40,9 @@ export default class LinkManager {
 
   getAllPathsWithOutgoingLinks(): string[] {
     const stopTimer = Logger.startTimer("LinkManager#getAllPathsWithOutgoingLinks");
-    const paths = [...this.#pathsWithKnownTypeMap.values()].filter((path) => this.getLinkedFilesFromPath(path).length);
+    const paths = [...this.#pathsWithKnownTypeMap.values()].filter(
+      (path) => this.getLinkedFilesFromPath(path).length,
+    );
     stopTimer();
     return paths;
   }
@@ -48,8 +50,21 @@ export default class LinkManager {
   /**
    * Determines whether the given path should be decorated as linked to the given file type and provides the decoration if so
    */
-  getFileTypeDecoratorData({ fileTypeName, path }: { fileTypeName: string; path: string }): DecorationData | undefined {
-    const endTimer = Logger.startTimer(`${fileTypeName}Decorator#getFileTypeDecoratorData: ${path} `);
+  getFileTypeDecoratorData({
+    fileTypeName,
+    path,
+  }: {
+    fileTypeName: string;
+    path: string;
+  }): DecorationData | undefined {
+    if (!this.pathIsRelevant(path)) {
+      return;
+    }
+
+    const endTimer = Logger.startTimer(
+      "Decorator#getFileTypeDecoratorData",
+      `${fileTypeName}-${path}`,
+    );
     try {
       const pathData = this.getPathData(path);
       if (!pathData) {
@@ -61,7 +76,9 @@ export default class LinkManager {
         (linkedFileType) => linkedFileType.name === fileTypeName,
       );
 
-      const outgoingLinksToDecorate = linkedDecoratorFileType?.getFilesMatching(pathData.file.pathKey);
+      const outgoingLinksToDecorate = linkedDecoratorFileType?.getFilesMatching(
+        pathData.file.pathKey,
+      );
       if (outgoingLinksToDecorate?.length) {
         return linkedDecoratorFileType?.getDecorationData(); // only apply allowed file type decoration if there are linked files of that type
       }
@@ -76,7 +93,7 @@ export default class LinkManager {
    * Gets all the files with outgoing links from the given path
    */
   getLinkedFilesFromPath(path: string): LinkedFileData[] {
-    const stopTimer = Logger.startTimer(`LinkManager#getLinkedFilesFromPath: ${path}`);
+    const stopTimer = Logger.startTimer("LinkManager#getLinkedFilesFromPath", path);
     try {
       const pathData = this.getPathData(path);
       if (!pathData) {
@@ -92,24 +109,33 @@ export default class LinkManager {
     }
   }
 
-  modifyFilesAndNotify({ removePaths = [], addPaths = [] }: { removePaths?: string[]; addPaths?: string[] }) {
+  modifyFilesAndNotify({
+    removePaths = [],
+    addPaths = [],
+  }: {
+    removePaths?: string[];
+    addPaths?: string[];
+  }) {
     Logger.info("#modifyFilesAndNotify", { removePaths, addPaths });
 
     const key = JSON.stringify({ removePaths, addPaths }, null, 2);
-    const endTimerAndLog = Logger.startTimer(`#modifyFilesAndNotify(${key})`);
+    const endTimerAndLog = Logger.startTimer("#modifyFilesAndNotify", key);
     try {
       removePaths = this.resolveDeletedPaths(removePaths);
       const relevantPathsRemoved = removePaths.filter((path) => this.pathIsRelevant(path));
       const relevantPathsAdded = addPaths.filter((path) => this.pathIsRelevant(path));
-      const relevantFilesAffected = relevantPathsRemoved.length || relevantPathsAdded.length;
-      if (!relevantFilesAffected) {
+      const relevantFilesAffectedCount = relevantPathsRemoved.length + relevantPathsAdded.length;
+      if (!relevantFilesAffectedCount) {
         return; // no known file types were affected so no need to update
       }
 
       this.applyPathRemovals(relevantPathsRemoved);
       this.applyPathAdditions(relevantPathsAdded);
 
-      const indirectlyAffectedPaths = this.getIndirectlyAffectedPaths([...relevantPathsAdded, ...relevantPathsRemoved]);
+      const indirectlyAffectedPaths = this.getIndirectlyAffectedPaths([
+        ...relevantPathsAdded,
+        ...relevantPathsRemoved,
+      ]);
       const affectedPaths = [...relevantPathsAdded, ...indirectlyAffectedPaths];
 
       Logger.info("#renameFilesAndNotify", {
@@ -132,7 +158,7 @@ export default class LinkManager {
     this.fileTypes.forEach((fileType) => fileType.dispose());
     this.fileTypes = [];
     this.#pathsWithKnownTypeMap.clear();
-    this.#pathDataCache.clear();
+    this.pathDataCache.clear();
   }
 
   /**
@@ -171,12 +197,20 @@ export default class LinkManager {
   }
 
   private applyPathAdditions(paths: Set<string> | string[]) {
+    const pathsCount = Array.isArray(paths) ? paths.length : paths.size;
+    if (!pathsCount) {
+      return;
+    }
+    Logger.info("#applyPathAdditions", paths);
     paths.forEach((path) => this.#pathsWithKnownTypeMap.set(normalisePath(path), path));
     this.fileTypes.forEach((fileType) => fileType.addPaths(paths));
   }
 
   private applyPathRemovals(paths: string[]) {
-    Logger.info("#applyPathRemovals", { resolvedPaths: paths });
+    if (!paths.length) {
+      return;
+    }
+    Logger.info("#applyPathRemovals", paths);
     paths.forEach((path) => this.#pathsWithKnownTypeMap.delete(normalisePath(path)));
     this.fileTypes.forEach((fileType) => fileType.removePaths(paths));
   }
@@ -198,8 +232,9 @@ export default class LinkManager {
    * so decided to simplify until there is a need for it
    */
   private getFileInfo(path: NormalisedPath): FileInfo | undefined {
-    if (!this.#pathsWithKnownTypeMap.has(path)) {
-      return; // we don't know about this path so we can't get any file info
+    Logger.count("LinkManager#getFileInfo CALLED");
+    if (!this.pathIsRelevant(path)) {
+      return;
     }
     for (const type of this.fileTypes) {
       const pathKey = type.getPathKeyFromPath(path);
@@ -210,21 +245,37 @@ export default class LinkManager {
   }
 
   private getIndirectlyAffectedPaths(pathsModified: string[]): string[] {
-    return pathsModified.flatMap((path) => this.getLinkedFilesFromPath(path)).map((file) => file.fullPath);
+    return pathsModified
+      .flatMap((path) => this.getLinkedFilesFromPath(path))
+      .map((file) => file.fullPath);
   }
 
+  /**
+   * This gets called multiple times for the same file for different file type decorators
+   * so effectiveness of cache depends on how many file types and files matching file types we have
+   * Keeping cache to optimise large codebases and to also minimise delay showing quick pick options
+   */
   private getPathData(path: string): PathData | null | undefined {
-    // this gets called multiple times for the same file for different file type decorators so caching makes sense
-    // also means we minimise delay showing quick pick options
+    Logger.count("LinkManager#getPathData CALLED");
     const normalisedPath = normalisePath(path);
-    const cached = this.#pathDataCache.get(normalisedPath);
+    const cached = this.pathDataCache.get(normalisedPath);
     if (typeof cached !== "undefined") {
+      Logger.count("LinkManager#getPathData CACHE HIT");
       return cached;
     }
+    Logger.count("LinkManager#getPathData CACHE MISS");
 
+    if (normalisedPath.endsWith("eslint/tests/lib/cli.js")) {
+      // debugger;
+    }
     const targetFile = this.getFileInfo(normalisedPath);
     if (!targetFile) {
-      this.#pathDataCache.set(normalisedPath, null);
+      Logger.count("LinkManager#getPathData: NO TARGET FILE");
+      /*
+      When a file is created VS Code tries to get its decorations before the file watcher has had a chance to add it to the list of known files
+      so a new file wont have file info here, so we dont do negative caching as it will be a hit next time
+      */
+      // this.pathDataCache.set(normalisedPath, null);
       return; // file is not of a known type
     }
 
@@ -234,7 +285,7 @@ export default class LinkManager {
       .filter((fileType) => fileType.allowsLinksFrom(targetFile.type));
 
     const metaData: PathData = { file: targetFile, allowedFileTypesForOutgoingLinks };
-    this.#pathDataCache.set(normalisedPath, metaData);
+    this.pathDataCache.set(normalisedPath, metaData);
     return metaData;
   }
 
@@ -246,7 +297,8 @@ export default class LinkManager {
   }
 
   /**
-   * @remark this gets called rarely, only when files are added or removed so doesn't need to be cached
+   * This gets called a lot but its not cached as it is called on initial load and when files/workspaces change
+   * which means the cache would be invalidated a lot
    */
   private pathIsRelevant(path: string): boolean {
     const shouldBeIgnored = this.ignorePatterns.some((regex) => regex.test(path));
